ai_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
ultralytics==8.0.222
opencv-python==4.8.1.78
torch==2.1.0
python-dotenv==1.0.0
numpy<2.0.0
